+++ src/QNEthernetServer.h
@@ -29,6 +29,11 @@ class EthernetServer final : public Server {
     return MEMP_NUM_TCP_PCB_LISTEN;
   }
 
+  // Return the port.
+  uint16_t port() const {
+    return port_;
+  }
+
   // Start listening on the server port. This calls begin(false).
   void begin() override;
 
@@ -55,7 +60,7 @@ class EthernetServer final : public Server {
   operator bool();
 
  private:
-  uint16_t port_;
+  const uint16_t port_;
   bool listening_;
 };